userReview/requirements.txt
async-generator==1.10
attrs==19.3.0
autopep8==1.4.3
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
billiard==3.5.0.4
bleach==3.1.3
blessed==1.16.1
celery==4.2.1
certifi==2020.4.5.1
cffi==1.13.2
codecov==2.0.15
colorama==0.3.9
coverage==4.5.1
cryptography==2.8
cssselect==1.0.3
cssutils==1.0.2
cx-Oracle==7.2.3
cycler==0.10.0
Cython==0.28.2
dask==0.19.4
dateutils==0.6.6
dbfread==2.0.7
deap==1.2.2
decorator==4.3.0
defusedxml==0.6.0
distributed==1.23.3
Django==2.2
django-cors-headers==3.2.0
django-discover-runner==1.0
django-elasticsearch-dsl==0.5.1
django-filter==1.1.0
django-gunicorn==0.1.1
django-htmlmin==0.10.0
django-picklefield==2.0
django-premailer==0.2.0
django-publish==1.0.5
django-q==1.0.2
django-redis==4.12.1
django-request-logging==0.7.0
django-rest-framework==0.1.0
django-rest-passwordreset==0.9.7
django-role-permissions==2.2.1
django-sendfile==0.3.11
django-silk==3.0.1
django-timezones==0.2
django-webpack-loader==0.5.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
future==0.16.0
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.17.3
hurry==1.0
hurry.filesize==0.9
idna==2.9
importlib-metadata==0.23
inflection==0.3.1
ipaddress==1.0.22
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.10
jmespath==0.9.4
joblib==0.14.1
json5==0.9.3
jsonfield==2.0.2
jsonschema==3.2.0
kiwisolver==1.0.1
kombu==4.2.1
lazy-object-proxy==1.3.1
leather==0.3.3
lxml==4.5.0
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
msgpack==0.5.6
nltk==3.3
notebook==6.0.3
openpyxl==3.0.3
outlier-utils==0.0.3
packaging==19.2
pandocfilters==1.4.2
paramiko==2.7.1
parsedatetime==2.5
parso==0.6.2
patsy==0.5.0
pbr==4.2.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.0
ply==3.11
premailer==3.2.0
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.4.7
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
py-translate==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.6.1
PyJWT==1.6.4
pylint==1.9.0
pymannkendall==1.2
PyNaCl==1.4.0
pyodbc==4.0.28
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.7
pytest==5.2.4
pytest-cov==2.6.0
pytest-django==3.4.4
python-dateutil==2.8.1
python-json-logger==0.1.10
python-logstash==0.4.6
python-pptx==0.6.11
python-slugify==4.0.0
pytimeparse==1.1.8
pytz==2019.3
PyYAML==3.13
pyzmq==19.0.0
requests==2.23.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rsa==4.0
saxpy==1.0.1.dev167
seasonal==0.3.1
Send2Trash==1.5.0
simple-salesforce==0.74.3
six==1.13.0
slimit==0.8.1
snakeviz==2.0.1
sortedcontainers==2.0.5
SQLAlchemy==1.3.15
sqlparse==0.3.0
stopit==1.1.2
tblib==1.3.2
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
timeout-decorator==0.4.1
toolz==0.9.0
TPOT==0.9.3
tqdm==4.23.3
traitlets==4.3.2
tsfresh==0.11.1
update-checker==0.16
uritemplate==3.0.1
urllib3==1.25.9
vine==1.1.4
voila==0.1.21
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.2.0
zict==0.1.3
zipp==0.6.0